requirements.txt
llama-index==0.10.30
disnake==2.9.1
python-dotenv==1.0.1
